+++ AutoAuctioneerAPI/Models/BidRequestModels/UpdateBidRequest.cs
@@ -8,5 +8,5 @@ public class UpdateBidRequest
 
     [Required] public Guid UserId { get; set; }
 
-    [Required] public int BidAmount { get; set; }
+    [Required] public long BidAmount { get; set; }
 }
