+++ DataAccessLayer_AutoAuctioneer/Models/Bid.cs
@@ -8,6 +8,6 @@ public class Bid
     public Guid Id { get; set; }
     public Guid ListingId { get; set; }
     public Guid UserId { get; set; }
-    public BigInteger BidAmount { get; set; }
+    public long BidAmount { get; set; }
     public DateTime? BidTime { get; set; }
 }
